fix(openlist): fork dev 升级后 onProcessExit 类型错（Phase 17 风险关闭）
**背景（Phase 17 风险登记）**
Hi-Sillot/OpenList@404daf0 dev 分支自带 openlistlib/event.go，
`OnProcessExit(code int)` 映射到 Java `onProcessExit(int)`。但
`OpenListBridge.kt:427` 写的是 `code: Long`——下一轮 AAR 重建必爆。

**根因**
- 写注释块时 Go→Java 类型映射认知不完整，以为是 `int64` → `Long`
- A2 fallback 里也是 `int64`——若 fork 哪天删了 event.go 触发 fallback，
  gomobile 也会生成 `long`，跟当前 Kotlin 不一致更糟

**修复（3 文件）**
1. `OpenListBridge.kt:427` `code: Long` → `code: Int`
2. `OpenListBridge.kt:392-403` 注释块列出 gomobile Go→Java 类型映射表
3. `build-openlist-aar.sh:390-400` A2 fallback `int64` → `int`，
   注释说明为何必须跟 fork 一致
4. `verification-discipline.md:348` Phase 17 风险登记 → Phase 21 关闭

**gomobile Go→Java 类型映射（已验证）**
- int / uint   → int (NOT long)
- int64 /uint64→ long
- int16 /uint16→ short
- string       → String

+++ app/encv-mobile/plugin-openlist/src/main/java/com/encvgo/plugin/openlist/OpenListBridge.kt
@@ -392,8 +392,15 @@ object OpenListBridge : Event, LogCallback {
 
     // === openlistlib.Event interface ===
     // gomobile generates **camelCase** Java method names from Go's PascalCase.
-    // Evidence: CI compile error lists abstract members as:
-    //   fun onProcessExit(p0: Long): Unit
+    // Evidence: read gomobile lowerFirst() (cmd/gobind/gen.go:527) + fork
+    // openlistlib/event.go @ Hi-Sillot/OpenList@404daf0. gomobile Go→Java type
+    // mapping for callback interfaces:
+    //   Go int   → Java int     (NOT long, even on 64-bit hosts)
+    //   Go int64 → Java long
+    //   Go int16 → Java short
+    //   Go string→ Java String
+    // Abstract members gomobile generates for `Event` + `LogCallback`:
+    //   fun onProcessExit(p0: Int): Unit   ← int, NOT long
     //   fun onShutdown(p0: String!): Unit
     //   fun onStartError(p0: String!, p1: String!): Unit
     //   fun onLog(p0: Short, p1: Long, p2: String!): Unit
@@ -417,7 +424,7 @@ object OpenListBridge : Event, LogCallback {
         broadcastStatus(0, false)
     }
 
-    override fun onProcessExit(code: Long) {
+    override fun onProcessExit(code: Int) {
         Log.e(TAG, "[SAT-DBG][OpenList] OnProcessExit() | code=$code | thread=${Thread.currentThread().name} | ts=${System.currentTimeMillis()}")
         val msg = "process exited with code $code"
         synchronized(lock) {
